% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign_where_available}
\alias{assign_where_available}
\title{Assign Data Tables to Connections Where Available}
\usage{
assign_where_available(available_dics, conns)
}
\arguments{
\item{available_dics}{A data frame containing the available data dictionaries.
The data frame should include columns \code{long_name}, \code{value}, and \code{cohort}.}

\item{conns}{A list of DataSHIELD connections indexed by cohort.}
}
\value{
None. This function is called for its side effects of assigning tables to connections.
}
\description{
This function takes a data frame of available data dictionaries and assigns tables to connections.
It uses the DataSHIELD function \code{datashield.assign.table} to assign each table to its respective cohort connection.
}
