% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server_up.R
\name{.check_server_up}
\alias{.check_server_up}
\title{Check if Server is Up}
\usage{
.check_server_up(login_data)
}
\arguments{
\item{login_data}{A data frame containing the login data with a column \code{url}.}
}
\value{
A tibble with columns \code{success} indicating if the server is up and \code{message} containing the error message if any.
}
\description{
This function checks if the servers specified in the login data are up and running.
It sends a request to each server URL and formats the error messages if any.
}
