% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{.count_vars}
\alias{.count_vars}
\title{Count Variables per Cohort}
\usage{
.count_vars(qc_stats)
}
\arguments{
\item{qc_stats}{A list containing QC statistics data. It is expected that \code{qc_stats} has a component named \code{variables}
which is a list of data frames. Each data frame should contain at least the columns \code{cohort} and \code{variable}.}
}
\value{
A tibble with two columns: \code{cohort} and \code{total_vars}. Each row represents a cohort and the number of distinct
variables associated with that cohort.
}
\description{
This function counts the number of distinct variables for each cohort in the provided QC statistics data.
}
