% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.format_dictionary_names}
\alias{.format_dictionary_names}
\title{Format Dictionary Names}
\usage{
.format_dictionary_names(files, neat_names)
}
\arguments{
\item{files}{A list of character vectors containing dictionary file names.}

\item{neat_names}{A list of neat names corresponding to the dictionary files.}
}
\value{
A data frame with formatted dictionary names.
}
\description{
This function formats dictionary names by creating a tidy data frame from the file names.
}
