% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.format_response}
\alias{.format_response}
\title{Extract Paths from API Response}
\usage{
.format_response(response)
}
\arguments{
\item{response}{A list containing the JSON response from the GitHub API.}
}
\value{
A character vector of file paths.
}
\description{
This function extracts file paths from the JSON response obtained from the GitHub API.
}
