% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{.get_all_columns}
\alias{.get_all_columns}
\title{Get All Columns}
\usage{
.get_all_columns(available_dics, conns)
}
\arguments{
\item{available_dics}{A data frame containing available dictionary information.}

\item{conns}{A datashield connections object}
}
\value{
A named list of column names for each dictionary.
}
\description{
This function retrieves all column names from the available dictionaries for each cohort.
}
