% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{.get_child_age_range}
\alias{.get_child_age_range}
\title{Get Child Age Range}
\usage{
.get_child_age_range(available_dics, repeated_dfs, conns)
}
\arguments{
\item{available_dics}{A data frame containing available dictionary information.}

\item{repeated_dfs}{A vector of data frames with repeated measures.}

\item{conns}{A datashield connections object}
}
\value{
A list of age statistics for children.
}
\description{
This function retrieves the age range of children from the repeated measures data.
}
