% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{.get_demographics}
\alias{.get_demographics}
\title{Get Demographics Data}
\usage{
.get_demographics(
  available_dics,
  core_df,
  yearly_df,
  core_vars,
  edu_var,
  repeated_dfs,
  conns
)
}
\arguments{
\item{available_dics}{A data frame containing available dictionary information.}

\item{core_df}{The core data frame to be used.}

\item{yearly_df}{The yearly data frame containing repeated measures.}

\item{core_vars}{A vector of core variables to be extracted.}

\item{edu_var}{The educational variable to be used.}

\item{repeated_dfs}{A vector of data frames with repeated measures.}
}
\value{
A list containing combined demographic statistics.
}
\description{
This function retrieves and processes demographic data from various data sources.
}
