% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.get_dics_from_api}
\alias{.get_dics_from_api}
\title{Retrieve Dictionary Paths from GitHub API}
\usage{
.get_dics_from_api()
}
\value{
A list containing the JSON response from the GitHub API.
}
\description{
This function makes a request to the GitHub API to get the list of dictionary files from the lifecycle-project's repository.
}
