% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.get_id_from_match}
\alias{.get_id_from_match}
\title{Get ID from Match}
\usage{
.get_id_from_match(ids, matches)
}
\arguments{
\item{ids}{A vector of IDs from reference dictionaries.}

\item{matches}{A logical vector indicating which entries in the reference dictionaries matched.}
}
\value{
A vector of matched IDs.
}
\description{
This function retrieves the ID corresponding to matched reference dictionary entries.
}
