% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{.get_max_sample}
\alias{.get_max_sample}
\title{Get Maximum Sample Size}
\usage{
.get_max_sample(non_rep_df, conns)
}
\arguments{
\item{non_rep_df}{The non-repeated measures data frame.}

\item{conns}{A datashield connections object}
}
\value{
A data frame with the maximum sample size for each cohort.
}
\description{
This function retrieves the maximum sample size for each cohort from a non-repeated measures data frame.
}
