% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{.get_oldest_youngest_ages}
\alias{.get_oldest_youngest_ages}
\title{Get Oldest and Youngest Ages}
\usage{
.get_oldest_youngest_ages(age_stats_combined, conns)
}
\arguments{
\item{age_stats_combined}{A data frame containing combined age statistics.}

\item{conns}{A datashield connections object}
}
\value{
A data frame with the oldest and youngest ages formatted.
}
\description{
This function retrieves the oldest and youngest ages from the combined age statistics.
}
