% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc_stats.R
\name{.get_qc_stats}
\alias{.get_qc_stats}
\title{Perform Quality Control for DataShield Connections}
\usage{
.get_qc_stats(login_data)
}
\arguments{
\item{login_data}{Tibble of server info.}
}
\value{
A list containing:
\item{packages}{Information about the DataShield packages available.}
\item{variables}{Statistics for the all variables.}
\item{demographics}{Specific demographic information.}
}
\description{
This function performs a series of quality control checks on DataShield connections. It fetches package information,
retrieves dictionary information, assigns variables, fetches variable statistics, and obtains key demographics.
}
