% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.get_stats_all_tables}
\alias{.get_stats_all_tables}
\title{Get Statistics for All Tables}
\usage{
.get_stats_all_tables(cohort_dics_available, conns)
}
\arguments{
\item{cohort_dics_available}{A data frame containing cohort dictionaries, including at least the columns 'cohort' and 'long_name'.}

\item{conns}{A list of connections, with names corresponding to the cohorts.}
}
\value{
A named list of statistics for each cohort.
}
\description{
This function retrieves statistics for a list of cohorts from a specified connection.
}
