% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{.get_stats_where_available}
\alias{.get_stats_where_available}
\title{Get Statistics Where Available}
\usage{
.get_stats_where_available(available_dics, filter_var, df, vars, conns)
}
\arguments{
\item{available_dics}{A data frame containing available dictionary information.}

\item{filter_var}{The variable used to filter the dictionary.}

\item{df}{The data frame from which to extract statistics.}

\item{vars}{A vector of variables to extract statistics for.}
}
\value{
A list of statistics for the specified variables.
}
\description{
This function retrieves statistics where available from the provided data sources.
}
