% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_qc.R
\name{.join_tokens_with_credentials}
\alias{.join_tokens_with_credentials}
\title{Join Tokens with Credentials}
\usage{
.join_tokens_with_credentials(credentials_split, token_values)
}
\arguments{
\item{credentials_split}{A list containing data frames of credentials split by server type. The list should have a named element \code{armadillo} which contains the Armadillo credentials.}

\item{token_values}{A data frame containing the token values to be joined with the Armadillo credentials. It should include a column \code{cohort} to join on.}
}
\value{
A data frame of Armadillo credentials with the token values joined.
}
\description{
This function joins the token values with the Armadillo credentials.
}
