% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.list_available_dics}
\alias{.list_available_dics}
\title{List Available Dictionaries}
\usage{
.list_available_dics(latest_version)
}
\arguments{
\item{latest_version}{The path to the latest version of the dictionary.}
}
\value{
A character vector of available dictionary files.
}
\description{
This function lists all available dictionary files in the latest version directory.
}
