% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.list_dictionary_types}
\alias{.list_dictionary_types}
\title{List Dictionary Types}
\usage{
.list_dictionary_types(path_to_dictionaries)
}
\arguments{
\item{path_to_dictionaries}{The path to the directory containing dictionaries.}
}
\value{
A character vector of paths to dictionary types.
}
\description{
This function lists all dictionary types available in the specified path.
}
