% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.list_latest_version}
\alias{.list_latest_version}
\title{List Latest Version}
\usage{
.list_latest_version(dict_types)
}
\arguments{
\item{dict_types}{A character vector of paths to dictionary types.}
}
\value{
A character string of the path to the latest version of the dictionary.
}
\description{
This function lists the latest version of a dictionary type.
}
