% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_qc.R
\name{.login_qc}
\alias{.login_qc}
\title{Login Quality Control Function}
\usage{
.login_qc(credentials)
}
\arguments{
\item{credentials}{A data frame containing the login credentials.
It should include columns like \code{cohort}, \code{url}, \code{server}, \code{username}, \code{password}, \code{value}.}
}
\value{
A list containing the successful connections and a summary of the login process.
}
\description{
This function handles the login process for both Armadillo and Opal servers.
It splits credentials by server, retrieves tokens for Armadillo servers,
builds login objects, handles the login process, and constructs a summary of the login process.
}
