% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.make_final_path_table}
\alias{.make_final_path_table}
\title{Create Final Path Table}
\usage{
.make_final_path_table(paths_as_tibble)
}
\arguments{
\item{paths_as_tibble}{A tibble with columns \code{type}, \code{version}, and \code{file}.}
}
\value{
A tibble with columns \code{type} and \code{name}.
}
\description{
This function processes the tibble to create a final table with unique dictionary names, removing misnamed dictionaries.
}
