% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{.make_first_year_from_repeats}
\alias{.make_first_year_from_repeats}
\title{Make First Year from Repeats}
\usage{
.make_first_year_from_repeats(available_dics, year_rep_df, edu_var, conns)
}
\arguments{
\item{available_dics}{A data frame containing available dictionary information.}

\item{year_rep_df}{The yearly data frame containing repeated measures.}

\item{edu_var}{The educational variable to be used.}

\item{conns}{A datashield connections object}
}
\value{
None. The function modifies the data in place.
}
\description{
This function processes repeated measures data to extract the earliest measure for each child.
}
