% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{.make_package_table}
\alias{.make_package_table}
\title{Create Package Table}
\usage{
.make_package_table(version_status)
}
\arguments{
\item{version_status}{A data frame where the row names represent package names.}
}
\value{
A tibble with a new column \code{Package} and the original columns of the input data frame.
}
\description{
This function converts the \code{version_status} data frame into a tibble with
the row names as a new column named \code{Package}.
}
