% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{.make_server_table}
\alias{.make_server_table}
\title{Create Server Table}
\usage{
.make_server_table(accessible)
}
\arguments{
\item{accessible}{A data frame with columns \code{success}, \code{message}, and \code{cohort}.}
}
\value{
A data frame with columns \code{type} and the cohorts, where \code{type} indicates
whether it is a success or error message.
}
\description{
This function processes the \code{accessible} data frame to create a table that
shows the success and error messages for different cohorts.
}
