% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{.make_table_headers}
\alias{.make_table_headers}
\title{Create Table Headers}
\usage{
.make_table_headers(table)
}
\arguments{
\item{table}{A data frame or tibble.}
}
\value{
A character vector with the headers for the table, where the first
element is an empty string.
}
\description{
This function generates a vector of table headers based on the column names
of the input table.
}
