% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{.rename_vars}
\alias{.rename_vars}
\title{Rename Variables in Transposed Data}
\usage{
.rename_vars(transposed)
}
\arguments{
\item{transposed}{A data frame containing a column named \code{measure}. The values in this column will be modified based on specific conditions.}
}
\value{
A data frame with the \code{measure} column values renamed:
"total_vars" is renamed to "Number of variables" and "max_obs" is renamed to "Maximum number of participants".
}
\description{
This function renames certain values in the \code{measure} column of the transposed data.
}
