% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.shorten_paths}
\alias{.shorten_paths}
\title{Filter and Shorten Paths}
\usage{
.shorten_paths(neat_response)
}
\arguments{
\item{neat_response}{A character vector of file paths.}
}
\value{
A character vector of shortened file paths.
}
\description{
This function filters the list of paths to include only \code{.xlsx} files and removes the \verb{dictionaries/} prefix.
}
