% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_qc.R
\name{.split_login_object}
\alias{.split_login_object}
\title{Split Login Object by Server}
\usage{
.split_login_object(builder)
}
\arguments{
\item{builder}{An object used to build the login data.}
}
\value{
A list of tibbles, each containing login data for a specific server.
}
\description{
This function builds the login object using the provided builder and then splits it by the \code{server} attribute.
}
