% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{.split_paths_to_tibble}
\alias{.split_paths_to_tibble}
\title{Split Paths into Tibble}
\usage{
.split_paths_to_tibble(shorter_paths)
}
\arguments{
\item{shorter_paths}{A character vector of shortened file paths.}
}
\value{
A tibble with columns \code{type}, \code{version}, and \code{file}.
}
\description{
This function splits the shortened paths into a tibble with separate columns for type, version, and file name.
}
