% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_dictionary_info}
\alias{format_dictionary_info}
\title{Format Dictionary Information}
\usage{
format_dictionary_info(coh_dics, all_dics)
}
\arguments{
\item{coh_dics}{A data frame containing cohort-specific dictionary information.}

\item{all_dics}{A data frame containing the master dictionary with all possible variables.}
}
\value{
A data frame with formatted dictionary information, showing the availability of variables for each cohort.
}
\description{
This function formats dictionary information by combining cohort-specific dictionaries with a master dictionary,
indicating the availability of each variable for each cohort, and restructuring the data for better readability.
}
