% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{identify_cohort_dics}
\alias{identify_cohort_dics}
\title{Identify Cohort Dictionaries}
\usage{
identify_cohort_dics(path_to_dic, conns)
}
\arguments{
\item{path_to_dic}{The path to the directory containing the dictionary files.}

\item{conns}{The connections object to the database where the data tables are stored.}
}
\value{
A final table containing cleaned and organized cohort dictionary information.
}
\description{
This function identifies cohort dictionaries by processing a directory of dictionary files
and then matches the extracted cohort dictionaries with the data tables. It cleans and organizes
the dictionary information and returns a final table.
}
