% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{list_all_dic_files}
\alias{list_all_dic_files}
\title{List All Dictionary Files}
\usage{
list_all_dic_files()
}
\value{
A data frame with all available dictionary files and their formatted names.
}
\description{
This function lists all dictionary files available in the specified path by combining all the helper functions.
}
