% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_qc_report}
\alias{make_qc_report}
\title{Generate a Quality Control Report}
\usage{
make_qc_report(qc_stats, sections, output_dir = NULL)
}
\arguments{
\item{qc_stats}{A list containing QC statistics.}

\item{sections}{Character, specifying which sections to include in the report. Options are "technical", "high-level" and "detailed"}

\item{output_dir}{A character string specifying the directory where the report will be saved. Defaults to the current working directory.}
}
\value{
The path to the generated report.
}
\description{
This function generates a quality control (QC) report from the provided QC statistics. It renders an RMarkdown document using the \code{qc-summary.Rmd} template and saves the report to the specified output directory.
}
