% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_qc.R
\name{perform_qc}
\alias{perform_qc}
\title{Perform Quality Control for DataShield Connections}
\usage{
perform_qc(conns)
}
\arguments{
\item{conns}{A list of DataShield connections.}
}
\value{
A list containing:
\item{packages}{Information about the DataShield packages available.}
\item{variables}{Statistics for the all variables.}
\item{demographics}{Specific demographic information.}
}
\description{
This function performs a series of quality control checks on DataShield connections. It fetches package information,
retrieves dictionary information, assigns variables, fetches variable statistics, and obtains key demographics.
}
