% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{summarise_demographics}
\alias{summarise_demographics}
\title{Summarize Demographics}
\usage{
summarise_demographics(variables, demographics)
}
\arguments{
\item{variables}{A list of data frames containing variable information.}

\item{demographics}{A data frame containing demographic information.}
}
\value{
A data frame summarizing demographic information.
}
\description{
This function summarizes demographic information based on the provided variables and demographics data.
}
