% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{summarise_obs_per_variable}
\alias{summarise_obs_per_variable}
\title{Summarize Observations per Variable}
\usage{
summarise_obs_per_variable(variables)
}
\arguments{
\item{variables}{A list of data frames containing variable information.}
}
\value{
A data frame summarizing the observations per variable for each cohort.
}
\description{
This function summarizes the observations per variable for each cohort.
}
