% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{summarise_total_vars_and_obs}
\alias{summarise_total_vars_and_obs}
\title{Summarize Variables}
\usage{
summarise_total_vars_and_obs(variables)
}
\arguments{
\item{variables}{A list of data frames containing variable information.}
}
\value{
A data frame summarizing the total number of variables and the maximum sample size for each cohort.
}
\description{
This function summarizes the variables available in the dictionaries by counting the total variables and
getting the maximum sample size for each cohort.
}
