% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{summarise_values_and_class}
\alias{summarise_values_and_class}
\title{Summarize Values and Class}
\usage{
summarise_values_and_class(variables)
}
\arguments{
\item{variables}{A list of data frames containing variable information.}
}
\value{
A data frame summarizing the values and class of the variables.
}
\description{
This function summarizes the values and class of the variables in the dictionaries.
}
