% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{summarise_variables}
\alias{summarise_variables}
\title{Summarize Variables}
\usage{
summarise_variables(qc_stats)
}
\arguments{
\item{qc_stats}{output from qc stats}
}
\value{
A data frame summarizing the total number of variables and the maximum sample size for each cohort.
}
\description{
This function summarizes the variables available in the dictionaries by counting the total variables and
getting the maximum sample size for each cohort.
}
